update(test/just.test): move into test dir, replace jest with qunit
